Remove the semi-useless DeepEqual.
Not only do the structs we get from upstream contain ServiceID and
Version field data that we don't have locally, but some upstream fields
are autofilled with defaults we don't know. For example, S3.Domain is
autoset to 's3.amazonaws.com' up in Fastly if we don't pass anything
when creating it.

Also, I believe there are some other Input things which won't quite
match, potentially regarding things that are Compatibooled.