Always sync all config options.
The previous method of checking for changes with the active version was
inadequate, as a previously created pending version may have had
previously prepared changes that we don't actually want to apply.

Turns out this is also a faster method as we don't do double listings of
everything, despite the fact that we're working on every possible config
object.