Fix the issue with empty dictionaries (#3)
* Fix the issue with empty dictionaries

Currently fastlyctl will try to reuse a previous version of the
service's config if there's one available. This means that there can be
cases where a service's version gets modified by fastlyctl when
validating a given change (even when in no-op mode) that interferes with
a later change.

This has hit us at least once that we're aware of:
https://reddit.atlassian.net/browse/SE-949

The proposed change is to lock every version after we're done validating
changes to the service. This will mean we'll stop reusing versions which
is likely to inflate the version count, but on the other hand it will
make this kind of bug impossible to reappear.